backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pillow>=10.0.0
boto3>=1.29.0
torch>=2.1.0
torchvision>=0.16.0
slowapi>=0.1.9
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiofiles>=23.2.0
firebase-admin>=6.2.0
